+++ examples/sandbox/sandbox.py
@@ -31,11 +31,17 @@ simConfig = specs.SimConfig()  # dictionary to store sets of simulation configur
 # NETWORK PARAMETERS
 ###############################################################################
 
+netParams.sizeX = 100
+netParams.sizeY = 50
+netParams.sizeZ = 20
+
+
+
 # Population parameters
-netParams.popParams['PYR1'] = {'cellModel': 'HH', 'cellType': 'PYR', 'numCells': 50} # pop of HH cells
-netParams.popParams['PYR2'] = {'cellModel': 'HH', 'cellType': 'PYR', 'numCells': 50} # pop of HH cells
+netParams.popParams['PYR1'] = {'cellModel': 'HH', 'cellType': 'PYR', 'gridSpacing': 10, 'xRange': [30,60]} # pop of HH cells
+netParams.popParams['PYR2'] = {'cellModel': 'HH', 'cellType': 'PYR', 'gridSpacing': 5, 'yRange': [20,40]} # pop of HH cells
 #netParams.popParams['artifVec'] = {'cellModel': 'VecStim', 'numCells': 10000, 'interval': 100, 'noise': 0.5, 'start': 50}  # pop of NetStims
-netParams.popParams['artif1'] = {'cellModel': 'VecStim', 'numCells': 100, 'rate': [0,5], 'noise': 1.0, 'start': 50}#, 
+# netParams.popParams['artif1'] = {'cellModel': 'VecStim', 'numCells': 100, 'rate': [0,5], 'noise': 1.0, 'start': 50}#, 
 #    'pulses': [{'start': 200, 'end': 300, 'rate': 50, 'noise':0.2}, {'start': 500, 'end': 800, 'rate': 30, 'noise':0.5}]}  # pop of NetStims
 
 # Synaptic mechanism parameters
@@ -85,7 +91,7 @@ netParams.addConnParams('artif1->PYR2',
 ###############################################################################
 
 # Simulation parameters
-simConfig.duration = 1*1e3 # Duration of the simulation, in ms
+simConfig.duration = 0.1*1e3 # Duration of the simulation, in ms
 simConfig.dt = 0.1 # Internal integration timestep to use
 simConfig.createNEURONObj = 1  # create HOC objects when instantiating network
 simConfig.createPyStruct = 1  # create Python structure (simulator-independent) when instantiating network
@@ -96,13 +102,14 @@ simConfig.recordTraces = {'Vsoma':{'sec':'soma','loc':0.5,'var':'v'}}
 
 # # Analysis and plotting 
 simConfig.analysis['plotRaster'] = True
+simConfig.analysis['plot2Dnet'] = True
 
 
 ###############################################################################
 # RUN SIM
 ###############################################################################
 
-#sim.createSimulateAnalyze()
-sim.create()
-sim.gatherData()
+sim.createSimulateAnalyze()
+#sim.create()
+#sim.gatherData()